refactor(combobox): split into 8 focused submodules
- Original combobox.rs: 1,646 lines -> new mod.rs: 261 lines (84% reduction)
- Created combobox/ directory with submodules:
  - option.rs - ComboOption type (55 lines)
  - state.rs - State getters (67 lines)
  - actions.rs - Action methods like open/close/select (123 lines)
  - input.rs - Input handling methods (169 lines)
  - filter.rs - Filtering logic (77 lines)
  - render.rs - View implementation (214 lines)
  - mod.rs - Main struct, builder methods, View trait impl (261 lines)
  - tests.rs - All 59 tests (727 lines)

All tests passing, clippy clean
